feat(maven-module) : Separate Java/Groovy sources to different modules
