+++ ContextAccess-core/src/main/java/org/bonitasoft/rest/context/RestContextAccessorJava.java
