+++ ContextAccess-modules/src/main/groovy/org/bonitasoft/rest/context/DocumentPermissionContextRule.groovy
